My team sees a need to improve performance of `from-db-object` for our production service.  Transients are a simple way to improve performance and, based on the project goals, I feel they can be justified:

- Target modern Clojure versions.
- Not carry technical debt from 2009 forever.

With transients I saw a ~30% uplift in performance according to criterium.

**BEFORE**:
```
Evaluation count : 1620 in 60 samples of 27 calls.
             Execution time mean : 37.643873 ms
    Execution time std-deviation : 749.896249 µs
   Execution time lower quantile : 36.990621 ms ( 2.5%)
   Execution time upper quantile : 38.966958 ms (97.5%)
                   Overhead used : 14.721958 ns

Found 4 outliers in 60 samples (6.6667 %)
	low-severe	 2 (3.3333 %)
	low-mild	 2 (3.3333 %)
 Variance from outliers : 7.8776 % Variance is slightly inflated by outliers

```

**AFTER**:
```
Evaluation count : 2400 in 60 samples of 40 calls.
             Execution time mean : 25.676598 ms
    Execution time std-deviation : 294.751226 µs
   Execution time lower quantile : 25.291556 ms ( 2.5%)
   Execution time upper quantile : 26.385678 ms (97.5%)
                   Overhead used : 14.721958 ns

Found 4 outliers in 60 samples (6.6667 %)
	low-severe	 3 (5.0000 %)
	low-mild	 1 (1.6667 %)
 Variance from outliers : 1.6389 % Variance is slightly inflated by outliers
```

Testing was performed by using [malli](https://github.com/metosin/malli) to generate nested maps and vectors, then converting those maps to `BasicDBObject/BasicDBList` types.  I used the same data for both runs.

The data I used: https://drive.google.com/file/d/1s1MiGtMrz_nk5ikcqkzOYUFT2ZRw35rg/view?usp=sharing

Here is the code I used to generate and run the benchmarks.

```clojure
(require '[malli.generator :as mg]
           '[criterium.core :as crit]
           '[monger.conversion :refer [from-db-object]])

  (import '[com.mongodb BasicDBObject BasicDBList])

  (defn convert [x]
    (cond
      (map? x) (reduce-kv
                (fn [a k v]
                  (.put a k (convert v))
                  a)
                (BasicDBObject.) x)
      (vector? x) (reduce-kv
                   (fn [a index v]
                     (.put a (str index) (convert v))
                     a)
                   (BasicDBList.) x)
      :else x))

  (def xs
    (as-> [:schema
           {:registry
            {::t [:or :int :nil :string]
             ::tt [:or
                   [:map-of :string [:or ::t [:ref ::tt]]]
                   [:vector [:or ::t [:ref ::tt]]]]}}
           [:ref ::tt]] $
      (mg/generator $)
      (mg/sample $ {:size 250})
      (map convert $)))

  (count xs)

  (crit/bench (mapv #(from-db-object % false) xs)) ;; run again with code changes
```